pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask==3.1.1",
    "lottie[all]==0.7.2",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
]
